[frontend] move formatDate function
