+++ frontend/src/lib/Functions.ts
@@ -1,0 +1,8 @@
+// format date to YYYY-MM-DD (for date input)
+export const formatDate = (dateString: string) => {
+  const date = new Date(dateString);
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, '0');
+  const day = date.getDate().toString().padStart(2, '0');
+  return `${year}-${month}-${day}`;
+};
