+++ frontend/src/pages/user/admin/allCoupons/newCoupon.tsx
@@ -1,6 +1,7 @@
 import { Row, Col } from 'react-bootstrap';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
+import { formatDate } from '@lib/Functions';
 import TButton from '@components/TButton';
 import FormItem from '@components/FormItem';
 import CouponItemTemplate from '@components/CouponItemTemplate';
@@ -15,14 +16,6 @@ interface IGlobalCouponDetail {
   start_date: string;
   type: 'percentage' | 'fixed' | 'shipping';
 }
-
-const formatDate = (dateString: string) => {
-  const date = new Date(dateString);
-  const year = date.getFullYear();
-  const month = (date.getMonth() + 1).toString().padStart(2, '0');
-  const day = date.getDate().toString().padStart(2, '0');
-  return `${year}-${month}-${day}`;
-};
 
 const NewAdminCoupon = () => {
   const navigate = useNavigate();
